Add a way to link in ADCIRC and DGSWEM libraries
The coupler can now be compiled with the following commands:
```
mkdir build
cmake \
    -DADCIRC_ROOT_DIR=<some_path>/adcirc-cg \
    -DDGSWEM_ROOT_DIR=<some_path>/dgswem \
    ..
make
```
(Also pass -DUSE_MPI=ON -DCMAKE_BUILD_TYPE=Release to CMake if needed)

The coupler links in the compiled **shared** libraries of ADCIRC and
DGSWEM, libadc.so and libdgswem.so. Therefore, users need to set the
LD_LIBRARY_PATH environment variable (Linux) to be able to run the
compiled binary. Obviously, the ADCIRC and DGSWEM shared libraries
must also be compiled first.